[IMP] Website tools made global
The truth is that these tools are quite useful outside of the website
module, as any other module can declare controllers or test them.

Thus, moving these tools to the tools namespace.

@Tecnativa TT32365
