+++ addons/website/models/mixins.py
@@ -8,12 +8,12 @@ from werkzeug.urls import url_join
 
 from odoo import api, fields, models, _
 from odoo.addons.http_routing.models.ir_http import url_for
-from odoo.addons.website.tools import text_from_html
 from odoo.http import request
 from odoo.osv import expression
 from odoo.exceptions import AccessError
 from odoo.tools import escape_psql
 from odoo.tools.json import scriptsafe as json_safe
+from odoo.tools.mail import text_from_html
 
 logger = logging.getLogger(__name__)
