+++ addons/website/models/website_page.py
@@ -5,10 +5,10 @@ from psycopg2 import sql
 import re
 
 from odoo.addons.http_routing.models.ir_http import slugify
-from odoo.addons.website.tools import text_from_html
 from odoo import api, fields, models
 from odoo.osv import expression
 from odoo.tools import escape_psql
+from odoo.tools.mail import text_from_html
 from odoo.tools.translate import _
