+++ addons/website/tests/test_fuzzy.py
@@ -7,9 +7,9 @@ import re
 from markupsafe import Markup
 
 from odoo.addons.website.controllers.main import Website
-from odoo.addons.website.tools import distance, MockRequest
 import odoo.tests
-from odoo.tests.common import TransactionCase
+from odoo.tests.common import MockRequest, TransactionCase
+from odoo.tools.misc import distance
 
 _logger = logging.getLogger(__name__)
