+++ addons/website_blog/models/website_blog.py
@@ -6,8 +6,8 @@ import random
 
 from odoo import api, models, fields, _
 from odoo.addons.http_routing.models.ir_http import slug, unslug
-from odoo.addons.website.tools import text_from_html
 from odoo.tools.json import scriptsafe as json_scriptsafe
+from odoo.tools.mail import text_from_html
 from odoo.tools.translate import html_translate
