+++ odoo/tools/misc.py
@@ -1736,3 +1736,62 @@ def get_flag(country_code: str) -> str:
     This emoji is composed of the two regional indicator emoji of the country code.
     """
     return "".join(chr(int(f"1f1{ord(c)+165:02x}", base=16)) for c in country_code)
+
+# Fuzzy matching tools
+
+def distance(s1="", s2="", limit=4):
+    """
+    Limited Levenshtein-ish distance (inspired from Apache text common)
+    Note: this does not return quick results for simple cases (empty string, equal strings)
+        those checks should be done outside loops that use this function.
+
+    :param s1: first string
+    :param s2: second string
+    :param limit: maximum distance to take into account, return -1 if exceeded
+
+    :return: number of character changes needed to transform s1 into s2 or -1 if this exceeds the limit
+    """
+    BIG = 100000  # never reached integer
+    if len(s1) > len(s2):
+        s1, s2 = s2, s1
+    l1 = len(s1)
+    l2 = len(s2)
+    if l2 - l1 > limit:
+        return -1
+    boundary = min(l1, limit) + 1
+    p = [i if i < boundary else BIG for i in range(0, l1 + 1)]
+    d = [BIG for _ in range(0, l1 + 1)]
+    for j in range(1, l2 + 1):
+        j2 = s2[j - 1]
+        d[0] = j
+        range_min = max(1, j - limit)
+        range_max = min(l1, j + limit)
+        if range_min > 1:
+            d[range_min - 1] = BIG
+        for i in range(range_min, range_max + 1):
+            if s1[i - 1] == j2:
+                d[i] = p[i - 1]
+            else:
+                d[i] = 1 + min(d[i - 1], p[i], p[i - 1])
+        p, d = d, p
+    return p[l1] if p[l1] <= limit else -1
+
+
+def similarity_score(s1, s2):
+    """
+    Computes a score that describes how much two strings are matching.
+
+    :param s1: first string
+    :param s2: second string
+
+    :return: float score, the higher the more similar
+        pairs returning non-positive scores should be considered non similar
+    """
+    dist = distance(s1, s2)
+    if dist == -1:
+        return -1
+    set1 = set(s1)
+    score = len(set1.intersection(s2)) / len(set1)
+    score -= dist / len(s1)
+    score -= len(set1.symmetric_difference(s2)) / (len(s1) + len(s2))
+    return score
